# whisper : fix VAD processing for skipped audio segments