@ggerganov 2025-06-06
> This caused the segment start timestamp to start at 0 but the end timestamp was relative to the skipped audio (so must larger and incorrect).

I think the end timestamp is correct. It is just the start timestamp that begins too early. But this is also what the reference OpenAI implementation produces, so I don't think it is a bug.

The proposed changes in this PR don't seem to work with `gb1.wav`:

```bash
make -j && ./build/bin/whisper-cli -f ./samples/gb1.wav -m models/ggml-medium.en.bin
```

```
main: processing './samples/gb1.wav' (3179750 samples, 198.7 sec), 1 threads, 1 processors, 5 beams + best of 5, lang = en, task = transcribe, timestamps = 1 ...

whisper_exp_compute_token_level_timestamps: audio samples skipped, setting segment.t0 to 2962

[00:00:29.620 --> 00:00:07.360]  [_BEG_] My fellow Americans, this day has brought terrible news and great sadness to our[_TT_368]
[00:00:07.360 --> 00:00:14.800]   country. At nine o'clock this morning, Mission Control in Houston lost contact[_TT_740]
[00:00:14.800 --> 00:00:21.400]   with our space shuttle Columbia. A short time later, debris was seen falling from[_TT_1070]
...
```