@danbev 2025-06-06
> I think the end timestamp is correct. It is just the start timestamp that begins too early. But this is also what the reference OpenAI implementation produces, so I don't think it is a bug.

This is what I'm seeing when I run this sample audio using OpenAI's whisper:
```console
whisper samples/aladdin-first30.mp3 --model medium.en --word_timestamps True
[00:20.040 --> 00:25.980]  There once lived a poor tailor, who had a son called Aladdin, a careless idle boy, who
[00:25.980 --> 00:29.920]  would do nothing but play all day long in the streets with little idle boys.
```
And the output in `aladdin-first30.json` looks like this:
```console
{
  "text": " There once lived a poor tailor, who had a son called Aladdin, a careless idle boy, who would do nothing but play all day long in the streets with little idle boys.",
  "segments": [
    {
      "id": 0,
      "seek": 0,
      "start": 20.039999999999996,
      "end": 25.98,
      "text": " There once lived a poor tailor, who had a son called Aladdin, a careless idle boy, who",
      "tokens": [
        50363,
        1318,
        1752,
        5615,
        257,
        3595,
        35280,
        11,
        508,
        550,
        257,
        3367,
        1444,
        978,
        46782,
        11,
        257,
        36138,
        21696,
        2933,
        11,
        508,
        51661
      ],
      "temperature": 0.0,
      "avg_logprob": -0.22081296920776367,
      "compression_ratio": 1.075,
      "no_speech_prob": 0.03549518063664436,
      "words": [
        {
          "word": " There",
          "start": 20.039999999999996,
          "end": 20.619999999999997,
          "probability": 0.18882536888122559
        },
        {
          "word": " once",
          "start": 20.619999999999997,
          "end": 21.2,
          "probability": 0.9850126504898071
        },
        {
          "word": " lived",
          "start": 21.2,
          "end": 21.5,
          "probability": 0.9961034059524536
        },
       ...
      {
          "word": " nothing",
          "start": 26.28,
          "end": 26.54,
          "probability": 0.9977497458457947
        },
        {
          "word": " but",
          "start": 26.54,
          "end": 26.82,
          "probability": 0.9585630893707275
        },
        {
          "word": " play",
          "start": 26.82,
          "end": 27.12,
          "probability": 0.9835618138313293
        },
        {
          "word": " all",
          "start": 27.12,
          "end": 27.32,
          "probability": 0.9963666200637817
        },
        {
          "word": " day",
          "start": 27.32,
          "end": 27.54,
          "probability": 0.9964172840118408
        },

```

If we only correct the the start timestamp (though that seems to be an incorrect fix I've made) we would get the following timestamps in the second segment:
```console
         "text": " nothing",                                                   
          "timestamps": {                                                       
            "from": "00:00:29,330",                                             
            "to": "00:00:30,020"                                                
          },                                                                    
          "offsets": {                                                          
            "from": 29330,                                                      
            "to": 30020                                                         
          },                                                                    
          "id": 2147,                                                           
          "p": 0.99731,                                                         
          "t_dtw": -1                                                           
        },                                                                      
        {                                                                       
          "text": " but",                                                       
          "timestamps": {                                                       
            "from": "00:00:30,040",                                             
            "to": "00:00:30,040"                                                
          },                                                                    
          "offsets": {                                                          
            "from": 30040,                                                      
            "to": 30040                                                         
          },                                                                    
          "id": 475,                                                            
          "p": 0.923541,                                                        
          "t_dtw": -1                                                           
        },                                                                      
        {                                                                       
          "text": " play",                                                      
          "timestamps": {                                                       
            "from": "00:00:30,040",                                             
            "to": "00:00:30,040"                                                
          },                                                                    
          "offsets": {                                                          
            "from": 30040,                                                      
            "to": 30040                                                         
          },                                                                    
          "id": 711,                                                            
          "p": 0.978036,                                                        
          "t_dtw": -1                                                           
        },                               

```
And actually this is also the case when running without first fix as well (so using the master branch). 
So I think there is migth be an issue with the initial start timestamp and these [repeating](https://github.com/danbev/learning-ai/blob/main/notes/whisper/skipped-audio-issue.md#root-cause) timestamps unless I'm missing something.